Encapsulate HoconConfigurationFactory implementation
